+++ src/Behat/Testwork/Cli/ServiceContainer/CliExtension.php
@@ -81,6 +81,7 @@ final class CliExtension implements Extension
     public function load(ContainerBuilder $container, array $config)
     {
         $this->loadCommand($container);
+        $this->loadSyntheticServices($container);
     }
 
     /**
@@ -101,6 +102,12 @@ final class CliExtension implements Extension
         $definition = new Definition('Behat\Testwork\Cli\Command', array('%cli.command.name%', array()));
         $definition->setPublic(true);
         $container->setDefinition(self::COMMAND_ID, $definition);
+    }
+
+    protected function loadSyntheticServices(ContainerBuilder $container)
+    {
+        $container->register(self::INPUT_ID)->setSynthetic(true)->setPublic(true);
+        $container->register(self::OUTPUT_ID)->setSynthetic(true)->setPublic(true);
     }
 
     /**